Got rid of bundles.db. Most operations in bundles/index.js are now synchronous as a result. Also added per-bundle npm functionality via "dependencies" nodecg.json prop that functions identially to npm's package.json.
For some reason, the _bundles var in bundles.index.js is being set to an empty array sometime after all bundles are parsed. This breaks the views and dashboard. I have no idea why this is happening and it's 3AM so I'm getting irrationally angry and emotional about this. Time for bed.