started messing around wiht making tests for the new functionality, and in the process made all the tests not work
I think the test files are all loaded by mocha at the same time. Then, all the "before" blocks are executed one after another. This sucks for us because we had written our test files assuming that the "after" of one test file would be run before the "before" of the next.