[replicants] avoid doubly applying changes
In cases where a replicant is declared both in a browser and an extension, changes executed in the browser would get applied twice. This is because the replicator would detect the change and apply it to its internal store of the data, but the extension-side replicant was still observing that data and reacting to the change. The replicator now emits an `unobserveReplicant` event to extensions before applying changes.