[replicants] Fix breakages caused by last commit.
Not only did I need to re-write some tests to account for this change, but changes that resulted in `path` being `[""]` would for some reason cause errors to get thrown. In those cases, I substitute `""` for the value of `path`. I have no idea why this works and the other value doesn't, but I ain't got time to worry about that shit.