[replicants] Always fire new "change" handlers
[ci skip]

When a new "change" listener is added, chances are that the developer wants it to be initialized ASAP. However, if the replicant has already been declared previously in the current content context, their "change" handler will *not* get run until another change comes in, which may never happen for Replicants that change very infrequently. To resolve this, we immediately invoke all new "change" handlers if appropriate.