[dashboard] Change how dialog buttons are handled.
Dialog buttons should now be declared in the package manifest. See https://gist.github.com/Lange/045460efbadfff7fc14e for an example.

Upon clicking a `dismiss` or `confirm` button, a `dialog-dismissed` or `dialog-confirmed` event will be emitted on the iframe's `document`, which code in the dialog can easily listen for.