feat(uploads): add file upload system
This adds an "Uploads" page to the side drawer panel on the dashboard,
accessible via the top left menu button. To activate the Uploads system
for a bundle, set `nodecg.uploads.enabled` to `true` in that bundle's `package.json`.
All bundles with Uploads enabled will appear in the Uploads page, where
files can be added, opened, and removed.

NodeCG will add an `uploads` Replicant to every bundle with Uploads
enabled. This Replicant is an Array that describes every file that has
been uploaded to that bundle. Each item in the array looks like this:

```
{
  sum: '0310a2d6c60ac389d2c846d9801cb717',
  base: 'printReceipt.ogg',
  ext: '.ogg',
  name: 'printReceipt',
  bundleName: 'lfg-sounds',
  url: '/uploads/lfg-sounds/printReceipt.ogg'
}
```

This Replicant is automatically updated whenever a file is added, changed,
or removed. The Uploads system will also react to changes made directly to
the folder, meaning that changes don't necessarily need to be made via the
"Uploads" page.

Closes #104