feat(dashboard): hide panel controls until mouseover
The three buttons present in the header of every dashboard panel were perhaps poorly designed. They got in the way of longer panel titles, and in short (1 width) panels, completely obscured the panel title.

This update hides these buttons by default, replacing them with a single left-pointing chevron that indicates to the user that there is something being hidden. When the user mouses over any part of the title bar, the buttons slide in from left and become usable as they were before.

This probably doesn't work whatsoever on mobile. I'll need to address that at some point.