feat(replicants): add `result` property to operations
Each object in the `operations` array emitted during Replicant `change` events now has a `result` property, which is the result of the method executed.

For instance, if the operation was a splice, this will contain the removed items.

This can be used to further improve NodeCG's interop with Polymer's data binding system.