fix(zeit_pkg): fix pkg'd executables not working when compiled on a drive other than C: (#498)
Turns out that the beginning drive letter in this path is inherited from the drive letter where the pkg was created. That is to say: if I invoke `pkg` from my `C:` drive, then the virtual filesystem snapshot will start with `c:`, if I invoke it from `D:`, then the virtual filesystem snapshot will start with `d:`, etc.

This was breaking my builds on Azure Pipelines, which often use a `D:` drive letter instead of `C:`.