fix: correct full update read code
readReplicant returns just the value of the replicant, not an object
containing the value and the revision number like this code expects,
resulting in replicants always being briefly set to undefined whenever
this code path is used.

Since this code path is only used in one instance where a revision
number mismatch happens, and we were told what revision number the
server actually has, it should be safe to use that for the assignment.