Target ES2015 in generated typings
The current build error on `main` is caused by the generated typings.
The used version of the `zod` dependency is using a private field which requires targeting
ES2015 in the generated typings otherwise the typescript compiler
will error causing the current build failure
Also the module resolution strategy needs to be specified when targeting
ES2015 which does PR also does.