Remove unused fields.
These fields were required by the TypeORM connector for express-session. However, we were setting the ttl for session to Infinity, effectively meaning that sessions would not expire. Since we need to build our own adapter to connect express-session to Drizzle, we can just omit these fields to get the same behavior. We should likely revisit this in the future, though, because this can cause the number of sessions we're storing to grow unbounded.